src/backend/requirements-api.txt
fastapi
uvicorn
python-multipart
numpy
torch
opencv-python
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
supabase
